+++ CHANGES.rst
@@ -27,6 +27,7 @@ AudioLazy changes history
 
 + general:
 
+  - Sphinx documentation
   - Integration with NumPy (tested on 1.5.0, 1.6.1 and 1.6.2) and MatPlotLib
     (tested on 1.0.1 and 1.2.0)
   - More docstrings and doctests