+++ CHANGES.rst
@@ -19,7 +19,7 @@
 AudioLazy changes history
 -------------------------
 
-*** Development... ***
+*** Version 0.04 (Documentation, LPC, Plots!) ***
 
 + examples: